pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mate-framework"
version = "0.1.0"
description = "A platform-agnostic Multi-Agent System framework for outdoor exploration."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.24.0",
    "aiosqlite>=0.19.0",
    "openai>=1.0.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "geopy>=2.4.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy"
]